Add the skeleton of the javascript and html/css
I'm sold on the design of the site so far and want to have as much
accessibility as possible in the long run so I'm trying to optimize
the site's functionality.

I want there to be a way to have notes and then I'm going to add an
Amazon affiliate link with the search of their book on Amazon.

Each entry will be edititable and able to be removed within the edit
menu.