Add hamburger menu and add menu; Set up book notes
Did a lot this commit. Added functioning menus. Each have similar looks,
but they're going to have different things in them. For now, I don't
remember what the hamburger menu is going to be used for, but I know I'm
planning on using the current add menu to edit additions in the future.

Clearly, I'm pleased with results and am looking forward to progress
with this website.

Planned:
    Ability to edit books added to the 'library'.
    Persistent storage functionality.
    Ability to remove books from the 'library'.