Add basic functionality to the webapp
I still don't know what the purpose of the 'menu' was for, but I've kept
it around for now. Books can now be successfully added, edited and
removed.

For now, I haven't tried the list with a large amount of 'Book's, but
it's functionable.

Next, I plan to add a 'reading progress' bar, so users can track how
much of the book they've read.