Add method to display book cover; Keep image's H:V
Added a new input for the user to include a URL of the book's cover.
Instead of making the image locked at 100px X 100px, I made it so the
image would keep it's aspect ratio and just added a margin at the top in
order to center the image.

Planned ideas:
    Persistent storage
    Login method to bring list with user